Changed float to double type, added default constructor to Cart class, added formating to printReceipt and added datat to Main class

+++ products/Beverage.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Beverage extends Perishable {
 
-	public Beverage(String name, String brand, float price, LocalDate expirationDate) {
+	public Beverage(String name, String brand, double price, LocalDate expirationDate) {
 		super(name, brand, price, expirationDate);
 		
 	}
