+++ products/Food.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Food extends Perishable {
 
-	public Food(String name, String brand, float price, LocalDate expirationDate) {
+	public Food(String name, String brand, double price, LocalDate expirationDate) {
 		super(name, brand, price, expirationDate);
 		
 	}
